Revert "Week1 #2496"
