+++ DuggaSys/diagram_figure.js
@@ -1,0 +1,6 @@
+/**
+ * Created by b15matki on 2017-03-30.
+ * lbl
+ */
+
+ // Declare three paths
