+++ DuggaSys/duggaed.js
@@ -83,6 +83,7 @@ function updateDugga()
 	var template=$("#template").val();
 	var release=$("#release").val();
 	var deadline=$("#deadline").val();
+	console.log(deadline);
 	
 	AJAXService("SAVDUGGA",{cid:querystring['cid'],qid:did,nme:nme,autograde:autograde,gradesys:gradesys,template:template,release:release,deadline:deadline,coursevers:querystring['coursevers']},"DUGGA");
 }
