+++ DuggaSys/profile.php
@@ -12,37 +12,31 @@ pdoConnect();
 	<meta name="viewport" content="width=device-width, initial-scale=1 maximum-scale=1">
 	<meta http-equiv="X-UA-Compatible" content="IE=edge">
 	<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-	<title>Section Editor</title>
+	<title>Profile</title>
 
-	
-	<!--<link type="text/css" href="../Shared/css/responsive.css" rel="stylesheet">-->
+	<link type="text/css" href="../Shared/css/style.css" rel="stylesheet">
 	<link type="text/css" href="../Shared/css/jquery-ui-1.10.4.min.css" rel="stylesheet">
 
 	<script src="../Shared/js/jquery-1.11.0.min.js"></script>
 	<script src="../Shared/js/jquery-ui-1.10.4.min.js"></script>
 	<script src="../Shared/dugga.js"></script>
-	<script src="contribution.js"></script>
-	<link type="text/css" href="../Shared/css/style.css" rel="stylesheet">
-	
-<!--the inline CSS was moved to the file shared/css/style.css-->
-
+	<script src="sectioned.js"></script>
 </head>
 <body>
 
 	<?php
-		$noup="COURSE";
+		$noup="PROFILE";
 		include '../Shared/navheader.php';
-	?>
-
-	<!-- content START -->
-	<div id="content">
-	
-	</div>
-	<!-- content END -->
-
-	<?php
 		include '../Shared/loginbox.php';
 	?>
+
+	<div id="content">
+		<form method="POST" action="">
+ 			Challenge question: <input onclick="saveSettings()" type="text" name="challengeQuestion" id="challengeQuestion" value="" />
+		 	<input type="button" name="Save" id="Save" value="Save" />
+		</form>
+		<button type="button" id="NotificationButton">Click here to activate notifications</button>
+	</div>
 
 </body>
 </html>
