--- Shared/SQL/addtemplates.sql
@@ -1,3 +1,0 @@
-insert into template (templateid,stylesheet,numbox) values(7,"template7.css",4);
-insert into template (templateid,stylesheet,numbox) values(8,"template8.css",3);
-insert into template (templateid,stylesheet,numbox) values(9,"template9.css",5);
