+++ Shared/dugga.js
@@ -11,17 +11,34 @@ var timeUsed;
 var stepsUsed;
 var inParams = "UNK";;
 var MAX_SUBMIT_LENGTH = 5000;
+var querystring=parseGet();
 
 function toggleloginnewpass(){
 
 	if(status == 0){
 		$("#newpassword").css("display", "block");
 		$("#login").css("display", "none");
+		$("#showsecurityquestion").css("display", "none");
+		$("#changepassword").css("display", "none");
 		status++;
 	}else if(status == 1){
 		$("#newpassword").css("display", "none");
 		$("#login").css("display", "block");
+		$("#showsecurityquestion").css("display", "none");
+		$("#changepassword").css("display", "none");
 		status= 0;
+	}else if(status == 2){
+		$("#newpassword").css("display", "none");
+		$("#login").css("display", "none");
+		$("#showsecurityquestion").css("display", "block");
+		$("#changepassword").css("display", "none");
+		status= 1;
+	}else if(status == 3){
+		$("#newpassword").css("display", "none");
+		$("#login").css("display", "none");
+		$("#showsecurityquestion").css("display", "none");
+		$("#changepassword").css("display", "block");
+		status= 1;
 	}
 
 }
@@ -138,11 +155,30 @@ function saveDuggaResult(citstr)
 		}
 		
 		AJAXService("SAVDU",{answer:citstr},"PDUGGA");
-
-		//alert("Kvitto - Duggasvar\n\n"+"\""+hexstr+"\"\n\nTeckensträngen ovan är ditt kvitto på att duggan har lämnats in.\n\nSpara kvittot på ett säkert ställe.");
+		
+		
+		
 		document.getElementById('receipt').value = hexstr;
-		document.getElementById('receiptInfo').innerHTML = "<p>\n\nTeckensträngen är ditt kvitto på att duggan har lämnats in. Spara kvittot på en säker plats.\n\n</p>";
+		
+		
+		var deadline = querystring['deadline']; //Get deadlinedate from URL
+		var currentTime = new Date().toJSON().slice(0,10).replace(/-/g,'-'); //Get current time
+		
+		if(deadline > currentTime){	//Check if deadline has past
+			
+			document.getElementById('receiptInfo').innerHTML = "<p>\n\nTeckensträngen är ditt kvitto på att duggan har lämnats in. Spara kvittot på en säker plats.\n\n</p>";
+		}
+		else{
+			
+			document.getElementById('receiptInfo').innerHTML = "<p>\n\nTeckensträngen är ditt kvitto på att duggan har lämnats in. Spara kvittot på en säker plats.\n\n</p><p>OBS! Du har lämnat in efter deadline. Läraren kommer att rätta dugga vid nästa kurstillfälle.";
+		
+		}
+		
+		//alert("Kvitto - Duggasvar\n\n"+"\""+hexstr+"\"\n\nTeckensträngen ovan är ditt kvitto på att duggan har lämnats in.\n\nSpara kvittot på ett säkert ställe.");
+		//}
+		
 		showReceiptPopup();
+	
 }
 
 //----------------------------------------------------------------------------------
@@ -353,10 +389,7 @@ function AJAXService(opt,apara,kind)
 				type: "POST",
 				data: "opt="+opt+para,
 				dataType: "json",
-				success: returnedAccess,
-				error: function(err) {
-					console.log("err:", err);
-				}
+				success: returnedAccess
 			});
 	}else if(kind=="SECTION"){
 			$.ajax({
@@ -448,6 +481,46 @@ function loginEventHandler(event){
 	}
 }
 
+
+
+function processResetPasswordCheckUsername() {
+
+	/*This function is supposed to get the security question from the database*/
+
+	var username = $("#newpassword #username").val();
+	
+	/*
+	$.ajax({
+			type:"POST",
+			url: "../Shared/resetpw.php",
+			data: {
+				username: username,
+				opt: "RESETPW"
+			},
+			var result = JSON.parse(data);
+			if(result['login'] == "success") {
+				$("#userName").html(result['username']);
+
+	*/
+
+	status = 2;
+	toggleloginnewpass();
+
+}
+
+function processResetPasswordCheckSecurityAnswer() {
+
+	/*This function is supposed to be resposible for checking so the sequrity question answer is correct and make the user go to the last step of changeing/recovering its password*/
+	
+	status = 3;
+	toggleloginnewpass();
+
+}
+
+function processResetPasswordChangePassword(){
+	/* This function is supposed to hande the change of the password if the two fucntions before it have proceeded without errors. 
+	Might be worth to add all the functions together (processResetPasswordCheckUsername, processResetPasswordCheckSecurityAnswer, processResetPasswordChangePassword)*/
+}
 
 function processLogin() {
