+++ Shared/navheader.php
@@ -59,7 +59,7 @@
 
 
 			if(checklogin()) {
-					echo "<td class='navName'><label id='userName' onclick='redirectToUMV()'>".$_SESSION['loginname']."</label></td>";
+					echo "<td class='navName'><a id='userName' href='profile.php'>".$_SESSION['loginname']."</a></td>";
 					echo "<td id='loginbutton' class='loggedin'><img id='loginbuttonIcon' onload='loginButtonHover(\"online\")' src='../Shared/icons/Man.svg' /></td>";
 			}else{
 					echo "<td class='navName'><label id='userName'>Guest</label></td>";
